+++ app/lib/ui/components/components.dart
@@ -21,11 +21,11 @@ import 'package:ods_flutter_demo/ui/components/material/component_material.dart'
 import 'package:ods_flutter_demo/ui/components/navigation_bar/navigation_bar.dart';
 import 'package:ods_flutter_demo/ui/components/progress/progress_circular.dart';
 import 'package:ods_flutter_demo/ui/components/progress/progress_linear.dart';
-import 'package:ods_flutter_demo/ui/components/radio_buttons/component_radio_buttons.dart';
+import 'package:ods_flutter_demo/ui/components/radio_buttons/radio_buttons.dart';
 import 'package:ods_flutter_demo/ui/components/sheets_bottom/sheets_bottom.dart';
 import 'package:ods_flutter_demo/ui/components/sliders/sliders.dart';
 import 'package:ods_flutter_demo/ui/components/snackbars/snackbars.dart';
-import 'package:ods_flutter_demo/ui/components/switches/component_switches.dart';
+import 'package:ods_flutter_demo/ui/components/switches/switches.dart';
 
 List<Component> components(BuildContext context) {
   return [